% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate.R
\docType{class}
\name{Candidate-class}
\alias{Candidate-class}
\title{A Candidate object}
\description{
Some more info on objects of the class \code{Candidate}, if you'd like
}
\details{
An object of the class 'Candidate' has the following slots:
\itemize{
\item \code{name} The candidate's name
\item \code{delegatesWon} Number of delegates the candidate has won to date
\item \code{party} Candidate's party
\item \code{delegatesNeeded} Number of delegates needed to secure party nomination
}
}
\author{
Matt Malis: \email{malis.matt@gmail.com}
}

