% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCandidate.R
\name{createCandidate}
\alias{createCandidate}
\alias{createCandidate,ANY-method}
\title{Creating a Candidate}
\usage{
createCandidate(name, delegatesWon, party)
}
\arguments{
\item{name}{A character object of the candidate's name}

\item{delegatesWon}{A numeric object of the delegates won by the candidate}

\item{party}{A character object of the candidate's party}
}
\value{
an object of class Candidate containing
\item{name}{The candidate's name}
\item{delegatesWon}{Number of delegates the candidate has won}
\item{party}{The candidate's party}
\item{delegatesNeeded}{Number of delegates needed for the candidate's party nomination}
}
\description{
Creates an object of class Candidate with user-provided inputs
}
\note{
This function creates a Candidate!
}
\examples{

cruzCandidate<-createCandidate("Ted Cruz", 135, "Republican")
}
\author{
Matt Malis \email{malis.matt@gmail.com}
}
\seealso{
\code{\link{totalNeeded}} \code{\link{propNeeded}}
}

