% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propNeeded.R
\name{propNeeded}
\alias{propNeeded}
\alias{propNeeded,ANY-method}
\title{Calculating proportion of delegates needed}
\usage{
propNeeded(candidate, delegatesRemaining)
}
\arguments{
\item{candidate}{An object of class Candidate}

\item{delegatesRemaining}{A numeric object of the number of delegates remaining}
}
\value{
A numeric object o the proportion of remaining delegates needed
}
\description{
Calculates proportion of remaining delegates needed to secure nomination
}
\note{
Don't have much else to say about this one
}
\examples{

propNeeded('Hillary Clinton', 2500)
}
\author{
Matt Malis \email{malis.matt@gmail.com}
}
\seealso{
\code{\link{createCandidate}}
}

