% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalNeeded.R
\name{totalNeeded}
\alias{totalNeeded}
\alias{totalNeeded,ANY-method}
\title{Calculating delegates needed}
\usage{
totalNeeded(party)
}
\arguments{
\item{party}{A character object of the candidate's party}
}
\value{
A numeric object of the number of delegates needed
}
\description{
Calculates total delegates needed to secure the candidate's party nomination
}
\note{
This function is called in the createCandidate function
}
\examples{

totalNeeded("Democrat")
}
\author{
Matt Malis \email{malis.matt@gmail.com}
}
\seealso{
\code{\link{createCandidate}}
}

